+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - Use [`$XDG_CACHE_HOME`](https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html) or `~/.cache` instead of `$XDG_CONFIG_HOME` or `~/config/` as base directory for API cache
 - Switch CI to use [setup-nextflow](https://github.com/nf-core/setup-nextflow) action to install Nextflow ([#1650](https://github.com/nf-core/tools/pull/1650))
 - Add tests for `nf-core modules update` and `ModulesJson`.
+- Add CI for GitLab remote [#1646](https://github.com/nf-core/tools/issues/1646)
+- Add `CITATION.cff` [#361](https://github.com/nf-core/tools/issues/361)
 - Allow customization of the `nf-core` pipeline template when using `nf-core create` ([#1548](https://github.com/nf-core/tools/issues/1548))
 
 ### Modules